+++ ios/Classes/handlers/MakePurchase/PurchasePromo.swift
@@ -15,20 +15,19 @@ final class PurchasePromoRequest: Request {
         if #available(iOS 12.2, *) {
             if Apphud.products() != nil {
                 guard let product = Apphud.product(productIdentifier: arguments.productId) else {
-                    result(FlutterError(code: "500", message: "product with id \(arguments.productId) does not exist", details: nil))
+                    result(["error" : "product with id \(arguments.productId) does not exist"])
                     return
                 }
                 Apphud.purchasePromo(product, discountID: arguments.discountID) { (response) in
-                    if (response.error != nil) {
-                        result(FlutterError(code: "500", message: response.error.debugDescription, details: nil))
-                    } else {
-                        result(response.toMap())
-                    }
+                    result(response.toMap())
                 }
             }
             else {
-                result(FlutterError(code: "500", message: "product with id \(arguments.productId) have not loaded yet or does not exist", details: nil))
+                result(["error" : "product with id \(arguments.productId) have not loaded yet or does not exist"])
             }
+        }
+        else {
+            result(["error" : "Apphud.purchasePromo is available only on iOS >= 12.2"])
         }
     }
 }