+++ lib/models/apphud_models/apphud_non_renewing_purchase.g.dart
@@ -10,7 +10,10 @@ ApphudNonRenewingPurchase _$ApphudNonRenewingPurchaseFromJson(Map json) {
   return ApphudNonRenewingPurchase(
     productId: json['productId'] as String,
     purchasedAt: (json['purchasedAt'] as num).toDouble(),
+    isActive: json['isActive'] as bool,
     canceledAt: (json['canceledAt'] as num?)?.toDouble(),
+    isLocal: json['isLocal'] as bool?,
+    isSandbox: json['isSandbox'] as bool?,
   );
 }
 
@@ -20,4 +23,7 @@ Map<String, dynamic> _$ApphudNonRenewingPurchaseToJson(
       'productId': instance.productId,
       'purchasedAt': instance.purchasedAt,
       'canceledAt': instance.canceledAt,
+      'isActive': instance.isActive,
+      'isSandbox': instance.isSandbox,
+      'isLocal': instance.isLocal,
     };